feat: waku chat hook and example chat component
The new files in this commit are used to work with the waku
network sockets servers. The chat functionalities can be used with the
hook 'useChat'. The business logic is decoupled from the component
logic, letting the component file focusing on styling and connections
with auth and etc.
Also an example chat component is commited.